feat: do stream tasks
