+++ src/streams/files/fileToWrite.txt
@@ -1,0 +1,3 @@
+sfojkfqw
+fkqlwfjqwf
+kqjfiqwjfiq
