+++ src/streams/transform.js
@@ -1,5 +1,12 @@
+import { Transform } from 'stream';
+
 const transform = async () => {
-    // Write your code here 
+  const reverseTransformer = Transform({
+    transform(chunk, encoding, callback) {
+      callback(null, chunk.toString().split('').reverse().join(''));
+    },
+  });
+  process.stdin.pipe(reverseTransformer).pipe(process.stdout);
 };
 
 await transform();
