+++ src/streams/write.js
@@ -1,5 +1,10 @@
+import { createWriteStream } from 'fs';
+
+const fileLoc = './files/fileToWrite.txt';
+
 const write = async () => {
-    // Write your code here 
+  const writableStream = createWriteStream(fileLoc);
+  process.stdin.on('data', data => writableStream.write(data));
 };
 
 await write();
